combine duplicate src folders and files from two repos
