+++ api/.env
@@ -1,0 +1,1 @@
+MONGO_URI=mongodb+srv://rwuser:9ngCkMV9Ur1zYAqo@cluster0.5efdq.mongodb.net/?retryWrites=true&w=majority
