+++ client/src/Styles/Styles.css
@@ -1,3 +1,18 @@
 @import 'tailwindcss/base';
 @import 'tailwindcss/components';
 @import 'tailwindcss/utilities';
+
+@layer utilities {
+    .grid-custom {
+        grid-template-colums: repeat(3, minmax(0, max-content));
+    }
+
+    .text-shadow {
+        text-shadow: 0px 1px 3px rgb(28, 33, 104) ;
+    }
+
+    .text-border {
+        -webkit-text-stroke: 0.5px white;
+    }
+
+}
