+++ client/src/components/Exams/Exams.js
@@ -1,5 +1,8 @@
 import { useState, useEffect } from "react";
 // import {Search} from "../Search/Search"
+import { Link } from "react-router-dom";
+import { Mobile } from "../Mobile/Mobile";
+import PopUp from "../PopUp/DataPopUp";
 
 // let data = require("../../data/exam-data.json");
 
@@ -10,10 +13,20 @@ import { useState, useEffect } from "react";
 // https://czi-covid-lypkrzry4q-uc.a.run.app/api/patient/COVID-19-AR-16424082
 
 export const Exams = ({ data, localData }) => {
+  // const data = useApi('exams');
+  // console.log('useApi: ', useApi('exams'));
 
+  const [data, setData] = useState([]);
   const [currentPage, setCurrentPage] = useState(1);
-  const [itemsPerPage, setItemsPerPage] = useState(5);
+  const [itemsPerPage, setItemsPerPage] = useState(4);
 
+  useEffect(() => {
+    fetch("https://czi-covid-lypkrzry4q-uc.a.run.app/api/exams")
+      .then((res) => res.json())
+      .then((res) => setData(res.exams))
+      .catch((error) => console.error("Error:", error));
+  }, []);
+  console.log(data)
 
   const handlePageChange = (page) => {
     setCurrentPage(page);
@@ -41,7 +54,57 @@ export const Exams = ({ data, localData }) => {
 
   return (
     <>
-      <div className="form flex items-center justify-center w-screen h-auto ">
+          {pageData.map(function (item) {
+        return (
+   <section className="md:flex relative w-full hidden px-4">
+            <div className="flex w-full justify-center p-4 md:max-w-7xl mx-auto h-auto border rounded-lg shadow-lg space-x-2">
+              <div className=" w-36">
+                <h1 className="font-bold text-base mb-1">Patient ID</h1>
+                <div>
+                  <Link><p className="text-sm text-blue-600 hover:font-bold hover:underline ">{item.patientId}</p></Link>
+                </div>
+              </div>
+              <div className=" w-28">
+                <h1 className="font-bold text-base mb-1">Exam ID</h1>
+                <Link><p className="text-sm text-blue-600 hover:font-bold hover:underline  ">{item.examId}</p></Link>
+              </div>
+              <div className=" w-36">
+              <h1 className="font-bold text-base mb-1">Images</h1>
+                <Link><img className="w-32" src={item.imageURL} alt="images"/></Link>
+              </div>
+              <div className=" w-80 px-2">
+                <h1 className="font-bold text-base mb-1">Key Findings</h1>
+                <p className="flex flex-wrap text-sm ">{item.keyFindings}</p>
+              </div>
+              <div className=" w-36">
+                <h1 className="font-bold text-base mb-1">Brixi Score</h1>
+                <p className="text-sm">{item.brixiaScores}</p>
+              </div>
+              <div className=" w-28 ">
+                <h1 className="font-bold text-base mb-1">Age</h1>
+                <p1 className="text-sm">{item.age}</p1>
+              </div>
+              <div className=" w-28 ">
+                <h1 className="font-bold text-base mb-1">Sex</h1>
+                <p className="text-sm">{item.sex}</p>
+              </div>
+              <div className=" w-28 ">
+                <h1 className="font-bold text-base mb-1">BMI</h1>
+                <p className="text-sm">{ item.bmi}</p>
+              </div>
+              <div className=" w-28 ">
+                <h1 className="font-bold text-base mb-1">Zip Code</h1>
+                <p className="text-sm">{ item.zipCode}</p>
+              </div>
+            </div>
+            <PopUp/>
+          </section>
+ );
+})}
+
+
+
+      {/* <div className="form flex items-center justify-center w-screen h-auto ">
         <div className=" m-auto p-6 h-screen">
           <div className="flex justify-center w-full p-6 bg-gray-200 border-black rounded-xl min-w-fit mt-4 ">
             <table className=" table-auto max-w-7xl min-w-fit text-left ">
@@ -96,7 +159,8 @@ export const Exams = ({ data, localData }) => {
             </div>
           </div>
         </div>
-      </div>
+      </div> */}
+            <Mobile/> 
     </>  
   );
 };
