--- src/components/Notfound/NotFound.js
@@ -1,10 +1,0 @@
-
-
-
-export const NotFound = () => {
-  return (
-    <div className="NotFound">
-      <p>Sorry. Something went wrong..</p>
-    </div>
-  );
-};
