--- src/components/Search/Search.js
@@ -1,20 +1,0 @@
-import React from 'react';
-
-export const Search = () => {
-    return (
-
-        <div className="flex justify-center container  mt-8">
-            <div className="flex  justify-around w-full md:w-2/5">
-                <input
-                    className="w-64 p-2 border border-gray-400 rounded"
-                    type="text"
-                    placeholder="Search exams"
-                />
-                <button className="bg-blue-500 p-2 rounded text-white">
-                    Search
-                </button>
-            </div>
-        </div>
-        
-    );
-};
