--- tailwind.config.js
@@ -1,10 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    "./src/**/*.{js,jsx,ts,tsx,css}",
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
